# phase1_openapi_v1.yaml
openapi: 3.0.3
info:
  title: Hybrid Fleet Platform API - Phase 1
  version: 1.0.0
  description: "Phase 1 OpenAPI draft aligned to SA / PRD / SD. Scope: Fleet Management Core + Dispatch Compliance Core."
servers:
  - url: https://api.example.com
    description: Production
  - url: https://sandbox-api.example.com
    description: Sandbox
security:
  - BearerAuth: []
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    TenantApiKey:
      type: apiKey
      in: header
      name: X-API-Key
    WebhookSignature:
      type: apiKey
      in: header
      name: X-Signature
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
          additionalProperties: true
        requestId:
          type: string
    Pagination:
      type: object
      properties:
        page:
          type: integer
          minimum: 1
        pageSize:
          type: integer
          minimum: 1
        totalItems:
          type: integer
          minimum: 0
        totalPages:
          type: integer
          minimum: 0
...
    Address:
      type: object
      properties:
        addressText:
          type: string
        lat:
          type: number
          format: double
        lng:
          type: number
          format: double
        placeId:
          type: string
        manualGeocode:
          type: boolean
          default: false
...
    Booking:
      type: object
      properties:
        bookingId:
          type: string
          format: uuid
        orderId:
          type: string
          format: uuid
        bookingType:
          type: string
          enum:
            - oneway
            - roundtrip
            - recurring
        reservationWindowStart:
          type: string
          format: date-time
        reservationWindowEnd:
          type: string
          format: date-time
        recurrenceRule:
          type: string
          nullable: true
        modifiableUntil:
          type: string
          format: date-time
          nullable: true
        cancelableUntil:
          type: string
          format: date-time
          nullable: true
    TenantBookingCreateRequest:
      type: object
      required:
        - scheduledAt
        - pickup
        - dropoff
        - passenger
        - bookingType
      properties:
        bookingType:
          type: string
          enum:
            - oneway
            - roundtrip
            - recurring
        scheduledAt:
          type: string
          format: date-time
        reservationWindowEnd:
          type: string
          format: date-time
        pickup:
          $ref: "#/components/schemas/Address"
        dropoff:
          $ref: "#/components/schemas/Address"
        passenger:
          oneOf:
            - type: object
              required:
                - passengerId
              properties:
                passengerId:
                  type: string
                  format: uuid
            - type: object
              required:
                - fullName
                - mobile
              properties:
                fullName:
                  type: string
                mobile:
                  type: string
                departmentName:
                  type: string
        vehicleForm:
          type: string
        costCenter:
          type: string
        partnerSubtype:
          type: string
          enum:
            - enterprise_dispatch
            - credit_card_airport_transfer
        onsiteContact:
          type: string
        flightRef:
          type: string
        note:
          type: string
          maxLength: 1000
...
paths:
...
  /api/tenant/bookings:
    post:
      tags:
        - TenantBookings
      summary: Create tenant booking
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TenantBookingCreateRequest"
      responses:
        "201":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    get:
      tags:
        - TenantBookings
      summary: List and filter tenant bookings with pagination and tenant isolation
      description: >-
        Queries authoritative tenant bookings with backend date, passenger, and status
        filtering, strict tenant isolation (anonymous 401, mismatched 403), stable sort
        (selected date DESC NULLS LAST, bookingId ASC), and authoritative filtered totals.
      parameters:
        - name: x-tenant-id
          in: header
          required: true
          description: Tenant selector. Real identity must match this selector.
          schema:
            type: string
        - name: q
          in: query
          required: false
          description: Trimmed case-insensitive literal search across passenger name, phone, email, bookingId, orderId (SQL wildcards escaped).
          schema:
            type: string
        - name: passenger
          in: query
          required: false
          description: Alias for passenger search text.
          schema:
            type: string
        - name: passengerId
          in: query
          required: false
          description: Canonical passenger UUID exact match.
          schema:
            type: string
        - name: status
          in: query
          required: false
          description: Booking status (active, completed, cancelled) or comma-separated list.
          schema:
            type: string
            enum:
              - active
              - completed
              - cancelled
        - name: bookingStatus
          in: query
          required: false
          description: Explicit booking status filter.
          schema:
            type: string
            enum:
              - active
              - completed
              - cancelled
        - name: fulfillmentStatus
          in: query
          required: false
          description: Order fulfillment status from OwnedOrderStatus union.
          schema:
            type: string
        - name: orderStatus
          in: query
          required: false
          description: Alias for fulfillmentStatus.
          schema:
            type: string
        - name: dateField
          in: query
          required: false
          description: Target date field for date range filtering and sorting. Default is reservationStart.
          schema:
            type: string
            enum:
              - reservationStart
              - createdAt
            default: reservationStart
        - name: dateFrom
          in: query
          required: false
          description: Start instant inclusive (explicit ISO 8601 with timezone, e.g. 2026-09-10T00:00:00Z).
          schema:
            type: string
            format: date-time
        - name: dateTo
          in: query
          required: false
          description: End instant exclusive (explicit ISO 8601 with timezone, e.g. 2026-09-11T00:00:00Z).
          schema:
            type: string
            format: date-time
        - name: page
          in: query
          required: false
          description: 1-based page number (positive integer >= 1). Default 1.
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: pageSize
          in: query
          required: false
          description: Page size (1 to 100). Default 20.
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: serviceBucket
          in: query
          required: false
          schema:
            type: string
        - name: subtype
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/Booking"
                  pagination:
                    $ref: "#/components/schemas/Pagination"
                  pageInfo:
                    $ref: "#/components/schemas/Pagination"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
...
  /api/tenant/audit:
    get:
      tags:
        - Audit
      summary: List tenant audit trail
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      type: object
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
...